refactor(report-issue): relocate component to view/components
Move dot-report-issue out of dot-toolbar/components since it is no
longer toolbar-specific. Update relative imports accordingly.
