+++ core-web/apps/dotcms-ui/src/app/view/components/dot-report-issue/dot-report-issue.component.html
