+++ core-web/apps/dotcms-ui/src/app/view/components/dot-toolbar/components/dot-toolbar-user/dot-toolbar-user.component.ts
@@ -17,7 +17,7 @@ import { DotToolbarUserStore } from './store/dot-toolbar-user.store';
 
 import { DotLoginAsComponent } from '../dot-login-as/dot-login-as.component';
 import { DotMyAccountComponent } from '../dot-my-account/dot-my-account.component';
-import { DotReportIssueComponent } from '../dot-report-issue/dot-report-issue.component';
+import { DotReportIssueComponent } from '../../../dot-report-issue/dot-report-issue.component';
 
 @Component({
     providers: [DotToolbarUserStore],
